requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
astroid==3.1.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
dill==0.3.8
Django==5.0.4
django-phone-verify==0.1.1
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
djangorestframework==3.15.1
frozenlist==1.4.1
idna==3.7
isort==5.13.2
mccabe==0.7.0
multidict==6.0.5
phonenumbers==8.13.35
phonenumberslite==8.13.35
platformdirs==4.2.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
pylint==3.1.0
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
tomlkit==0.12.4
twilio==9.0.5
urllib3==2.2.1
yarl==1.9.4
